pyproject.toml
[project]
name = "semantic-kernel-workshop"
version = "0.1.0"
description = "A workshop for Semantic Kernel"
readme = "README.md"
requires-python = "==3.12.*"
dependencies = [
    "azure-monitor-opentelemetry>=1.6.10",
    "fastmcp>=2.2.0",
    "jupyter>=1.1.1",
    "jupyterlab>=4.5.0a0",
    "mermaid-py>=0.7.1",
    "openai>=1.90.0",
    "python-dotenv==1.0.1",
    "rich>=14.0.0",
    "semantic-kernel[azure]==1.31.0",
    "tzdata>=2025.2",
]
